modules/authorizer/requirements.txt
boto3==1.24.91
PyJWT==2.3.0